requirements.txt
# Wankr Bot — xAI (Grok); langchain-openai speaks OpenAI-compatible API (api.x.ai)
pandas>=2.0
langchain>=1.2
langchain-openai>=0.2
langchain-core>=1.2
langchain-community>=0.4
langchain-classic>=1.0
llama-index>=0.14
openai>=1.0
tweepy>=4.14
requests>=2.28
python-dotenv>=1.0
flask>=3.0
flask-cors>=4.0
infisicalsdk>=1.0
